+++ bw.core/src/main/java/sa/gov/nic/bio/bw/core/utils/GuiUtils.java
@@ -786,6 +786,8 @@ public class GuiUtils implements AppLogger
 		
 		fingerprintBase64Images.forEach((position, fingerprintImage) ->
 		{
+			if(fingerprintImage == null) return;
+
 			ImageView imageView = imageViewMap.get(position);
 		    String dialogTitle = dialogTitleMap.get(position);
 		
@@ -1124,5 +1126,13 @@ public class GuiUtils implements AppLogger
 			}
 		}
 		return null;
+	}
+
+	public static String getPageCounterFooterInArabic(int pageNumber, int pagesCount)
+	{
+		ResourceBundle resourceBundle = AppUtils.getCoreStringsResourceBundle(AppConstants.Locales.SAUDI_AR_LOCALE);
+		String text = resourceBundle.getString("label.thePage") + " " + pageNumber + " " +
+														resourceBundle.getString("label.from") + " " + pagesCount;
+		return AppUtils.localizeNumbers(text, AppConstants.Locales.SAUDI_AR_LOCALE, false);
 	}
 }